Change interpretation of the graph parameter in the N3Store
Setting the `graph` parameter to a falsy value will now,
just like with the other parameters, be interpreted as a wildcard.
For fetching triples within the default graph,
this graph explicitly needs to be provided as parameter value.
This default graph can be found in the store in the read-only field `defaultGraph`.
This can be modified by providing a defaultGraph field in the constructor of the N3Store.