BREAKING CHANGE: Make class Term constructors internal
These constructors take the `id` at face value so can be interpreted as the
wrong term type if an invalid value is given; for instance a `NamedNode` can
be interpred as a `Literal` if instantiated with `new NamedNode('"hellow world"')`.

Consumers of the library should use the `DataFactory` instead to create new terms.